remove count-related methods and keys and modify gateway id type to uint8

+++ app/v1/asset/gateway.go
@@ -6,7 +6,7 @@ import (
 
 // Gateway represents the gateway
 type Gateway struct {
-	ID         uint64           `json:"id"`             //  ID of the gateway
+	ID         uint8            `json:"id"`             //  ID of the gateway
 	Owner      sdk.AccAddress   `json:"owner"`          //  Owner address of the gateway
 	Identity   string           `json:"identity"`       //  Identity of the gateway
 	Moniker    string           `json:"moniker"`        //  Moniker of the gateway
